+++ lib/components/base/BaseDivider.vue
@@ -9,6 +9,9 @@ hr(
     "dm-base-divider--" + size,
     "dm-base-divider--" + color
   ]`
+  :style=`{
+    margin: computedMargin
+  }`
 )
 </template>
 
@@ -37,11 +40,29 @@ export default {
         );
       }
     },
+    margin: {
+      type: String,
+      default: null
+    },
     size: {
       type: String,
       default: "large",
       validator(x) {
         return ["small", "large"].indexOf(x) !== -1;
+      }
+    }
+  },
+
+  computed: {
+    computedMargin() {
+      if (this.margin) {
+        return this.margin;
+      }
+
+      if (this.size === "small") {
+        return "20px auto";
+      } else if (this.size === "large") {
+        return "40px auto";
       }
     }
   }
@@ -76,14 +97,12 @@ $colors: black, blue, green, orange, purple, red, turquoise, white;
   // --> SIZES <--
 
   &--small {
-    margin: 20px auto;
     width: 60px;
     height: 4px;
     border-top-width: 4px;
   }
 
   &--large {
-    margin: 40px auto;
     width: 100%;
     height: 1px;
     border-top-width: 1px;